refactor(log): extract logging functionality to logging.rs
Move Log trait, Logger struct, LogCategory and LogLevel enums with their
From implementations, callback functions, and disable_logging function
to dedicated logging.rs module.
