+++ src/log/mod.rs
@@ -1,0 +1,3 @@
+pub mod logging;
+
+pub use logging::{disable_logging, Log, LogCategory, LogLevel, Logger};
